% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_colors.R
\name{bird_colors}
\alias{bird_colors}
\title{Extract a bird's color palette}
\usage{
bird_colors(
  palette_name = "Scarlet Macaw",
  ncols = NA,
  reverse = FALSE,
  expand_palette = FALSE
)
}
\arguments{
\item{palette_name}{Common name of a bird}

\item{ncols}{Number of colors to extract}

\item{reverse}{Should the palette order be flipped?}

\item{expand_palette}{Should the palette be expanded to more colors?}
}
\value{
Extracted bird colors.
}
\description{
Extract a bird's color palette
}
\examples{
bird_colors()
bird_colors("Scarlet Macaw")
bird_colors("Scarlet_Macaw",7)
bird_colors("Scarlet_Macaw",4,reverse=TRUE)
bird_colors("Bluethroat",4,reverse=TRUE,expand_palette=15)

}
