% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_colors.R
\name{bird_menu}
\alias{bird_menu}
\title{Return available bird color palettes}
\usage{
bird_menu(all_or_rec = "all")
}
\arguments{
\item{all_or_rec}{Option "all" returns all available palettes, "rec" returns recommended}
}
\value{
Returns data.frame of available birds and color number
}
\description{
Return available bird color palettes
}
\examples{
bird_menu()
}
