% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_colors.R
\name{bird_palette_visualizer}
\alias{bird_palette_visualizer}
\title{Print out pictures of available bird color palettes}
\usage{
bird_palette_visualizer(all_or_rec = "all", pdf_plot = FALSE)
}
\arguments{
\item{all_or_rec}{Option "all" returns all available palettes, "rec" returns recommended}

\item{pdf_plot}{Create a pdf of available palettes in the current working directory. Default is to send directly to plots}
}
\value{
Creates pdf of all available birds or rec birds
}
\description{
Print out pictures of available bird color palettes
}
\examples{
bird_palette_visualizer("all")
}
