% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_colors.R
\name{bird_search}
\alias{bird_search}
\title{Search for available bird color palettes}
\usage{
bird_search(sp_name = "Penguin")
}
\arguments{
\item{sp_name}{Provide full or partial name of bird you are looking for}
}
\value{
Returns character string of available birds
}
\description{
Search for available bird color palettes
}
\examples{
bird_search("Hummingbird")
}
