% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_colors.R
\name{leaderboard}
\alias{leaderboard}
\title{Return leaders in contributing bird color palettes}
\usage{
leaderboard()
}
\value{
Returns data.frame of leaderboard in contributing bird color palettes
}
\description{
Return leaders in contributing bird color palettes
}
\examples{
leaderboard()
}
