% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_colors.R
\name{scale_color_bird}
\alias{scale_color_bird}
\title{Create ggplot color gradient}
\usage{
scale_color_bird(bird_cols, midpoint = NA)
}
\arguments{
\item{bird_cols}{Bird colors extracted via bird_colors function}

\item{midpoint}{Midpoint of diverging color palette, only for 3-color palettes}
}
\value{
ggplot-friendly color gradient
}
\description{
Create ggplot color gradient
}
\examples{
scale_color_bird(bird_colors())

}
