% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_colors.R
\name{bird_colors}
\alias{bird_colors}
\title{Extract a bird's color palette}
\usage{
bird_colors(palette_name = "Scarlet Macaw", ncols = NA)
}
\arguments{
\item{palette_name}{Common name of a bird}

\item{ncols}{Number of colors to extract}
}
\value{
Extracted bird colors.
}
\description{
Extract a bird's color palette
}
\examples{
bird_colors()
bird_colors("Scarlet Macaw")
bird_colors("Scarlet_Macaw",7)

}
