% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_colors.R
\name{bird_menu}
\alias{bird_menu}
\title{Print out available bird color palettes}
\usage{
bird_menu()
}
\value{
Prints data.frame of available birds and color number
}
\description{
Print out available bird color palettes
}
\examples{
bird_menu()
}
