% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_colors.R
\name{scale_color_bird}
\alias{scale_color_bird}
\title{Create ggplot color gradient}
\usage{
scale_color_bird(bird_cols)
}
\arguments{
\item{bird_cols}{Bird colors extracted via bird_colors function}
}
\value{
ggplot-friendly color gradient
}
\description{
Create ggplot color gradient
}
\examples{
scale_color_bird(bird_cols)

}
