% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_colors.R
\name{scale_fill_bird}
\alias{scale_fill_bird}
\title{Create ggplot fill gradient}
\usage{
scale_fill_bird(bird_cols, midpoint = NA)
}
\arguments{
\item{bird_cols}{Bird colors extracted via bird_colors function}

\item{midpoint}{Midpoint of diverging color palette, only for 3-color palettes}
}
\value{
ggplot-friendly fill gradient
}
\description{
Create ggplot fill gradient
}
\examples{
scale_fill_bird(bird_colors())

}
