% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_colors.R
\name{scale_fill_bird}
\alias{scale_fill_bird}
\title{Create ggplot fill gradient}
\usage{
scale_fill_bird(bird_cols)
}
\arguments{
\item{bird_cols}{Bird colors extracted via bird_colors function}
}
\value{
ggplot-friendly fill gradient
}
\description{
Create ggplot fill gradient
}
\examples{
scale_fill_bird(bird_colors())

}
